Introduced LightweightPattern[Enter|Exit]Event. This is part of an attempt to be able to detect the navigation along an "each time"-edges in case of "forEach"-nodes as well as the returning edges to the "forEach"-node.
Rationale: Unfortunately, technical restrictions exist that prevent the correct detection of the outgoing controlflow and the incoming controlflow for "forEach"-nodes. For example, it is not possible to determine whether an activity node is the "last" node before returning control to the "forEach"-node during the process of code generation (at least not within the bounds defined by reasonable effort). It is also not possible to issue ordinary Pattern[Enter|Exit]Events, because there exists no place in the CodeGen2 templates that explicitly merge the code for the adjacent nodes (those nodes that are part of the flow of control started by the "each time"-edge) with the "inner"/contained code produced for the "forEach"-pattern itself. (This merging is done implicitly by a corresponding java-class that is part of the controller framework. Manipulating the java class is not a good idea, since this code has beend generated itself and tampering with the code includes the high risk of breaking the complete code generator.)